requirements.txt
mysql-connector-python==9.0.0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
